app/requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
pyarrow
mysql-connector-python
